Added shared Vuforia key + executable op-modes

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pixelanalysis/HSV.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.pixelanalysis;
 
 public class HSV {
-    public double h;
-    public double s;
-    public double v;
+    public int h;
+    public int s;
+    public int v;
 }
